refactor: extract the return-type cluster into SteepBridge::ReturnTypeAnalyzer
`method_return_types_by_kind` and the helpers only it uses —
`singleton_class_def_ids`, `resolve_block_generic_type`, `BLOCK_GENERIC_METHODS`
and the `block_mismatches` indexing — move into a collaborator, following the
shape `IvarWriteAnalyzer` and `BlockAnalyzer` already set.

`singleton_class_def_ids` and `resolve_block_generic_type` were public on the
bridge with exactly one caller each (the method that moved), so they become
private: the bridge's public surface drops 21 → 19. steep_bridge.rb 773 → 659.

The class comment states the split against `Inference::ReturnTypeResolver`,
which consumes this and is the one that decides the final return type.
