+++ lib/rbs_infer/analyzer.rb
@@ -1270,5 +1270,6 @@ require_relative "signatures/steep_bridge/lexical_scope"
 require_relative "signatures/steep_bridge/type_formatter"
 require_relative "signatures/steep_bridge/ivar_write_analyzer"
 require_relative "signatures/steep_bridge/block_analyzer"
+require_relative "signatures/steep_bridge/return_type_analyzer"
 require_relative "project/source_expanders"
 require_relative "project/self_type_annotators"
